Reorganization (Python purists will hate me).
